api/floyd_requirements.txt
numpy
flask
torch
torchvision
matplotlib
scipy
opencv-python
keras
Pillow